refactor: major WebSocket architecture improvements per code review
- Extract WebSocketManager responsibilities into focused services:
  - WebSocketConnectionService for connection lifecycle management
  - WebSocketAuthService for authentication and authorization
  - WebSocketMessagingService for unified messaging operations
- Add queue size limits when circuit breaker is open
- Improve latency calculation for low sample counts
- Create custom ExponentialBackoffException for better error handling
- Fix priority queue starvation with round-robin processing
- Centralize channel parsing logic in auth service
- Consolidate duplicate send_to_* logic into messaging service
- Fix 7 tests (from 15 failed to 8 failed)

This refactoring improves maintainability, testability, and reliability
